basic entry and retrieval of weather data implemented
