+++ WeatherApp/Pages/WeatherView/Create.cshtml
@@ -1,5 +1,5 @@
 @page
-@model WeatherApp.Pages.Weather.CreateModel
+@model WeatherApp.Pages.WeatherView.CreateModel
 
 @{
     ViewData["Title"] = "Create";
