+++ WeatherApp/Pages/WeatherView/Delete.cshtml.cs
@@ -7,17 +7,18 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 using WeatherApp.Models;
 
-namespace WeatherApp.Pages.Weather
+namespace WeatherApp.Pages.WeatherView
 {
-    public class DetailsModel : PageModel
+    public class DeleteModel : PageModel
     {
         private readonly WeatherApp.Models.WeatherContext _context;
 
-        public DetailsModel(WeatherApp.Models.WeatherContext context)
+        public DeleteModel(WeatherApp.Models.WeatherContext context)
         {
             _context = context;
         }
 
+        [BindProperty]
         public Weather Weather { get; set; }
 
         public async Task<IActionResult> OnGetAsync(int? id)
@@ -34,6 +35,24 @@ namespace WeatherApp.Pages.Weather
                 return NotFound();
             }
             return Page();
+        }
+
+        public async Task<IActionResult> OnPostAsync(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            Weather = await _context.Weather.FindAsync(id);
+
+            if (Weather != null)
+            {
+                _context.Weather.Remove(Weather);
+                await _context.SaveChangesAsync();
+            }
+
+            return RedirectToPage("./Index");
         }
     }
 }
