+++ WeatherApp/Pages/WeatherView/Details.cshtml
@@ -1,13 +1,12 @@
 @page
-@model WeatherApp.Pages.Weather.DeleteModel
+@model WeatherApp.Pages.WeatherView.DetailsModel
 
 @{
-    ViewData["Title"] = "Delete";
+    ViewData["Title"] = "Details";
 }
 
-<h2>Delete</h2>
+<h2>Details</h2>
 
-<h3>Are you sure you want to delete this?</h3>
 <div>
     <h4>Weather</h4>
     <hr />
@@ -61,13 +60,8 @@
             @Html.DisplayFor(model => model.Weather.UVIndex)
         </dd>
     </dl>
-    
-    <form method="post">
-        <input type="hidden" asp-for="Weather.ZIP" />
-        <input type="hidden" asp-for="Weather.City" />
-        <input type="hidden" asp-for="Weather.State" />
-        <input type="hidden" asp-for="Weather.Day" />
-        <input type="submit" value="Delete" class="btn btn-default" /> |
-        <a asp-page="./Index">Back to List</a>
-    </form>
+</div>
+<div>
+    @Html.ActionLink("Edit", "Edit", new { /* id = Model.PrimaryKey */ }) |
+    <a asp-page="./Index">Back to List</a>
 </div>
