+++ WeatherApp/Pages/WeatherView/Edit.cshtml
@@ -1,5 +1,5 @@
 @page
-@model WeatherApp.Pages.Weather.EditModel
+@model WeatherApp.Pages.WeatherView.EditModel
 
 @{
     ViewData["Title"] = "Edit";
