+++ WeatherApp/Pages/WeatherView/Index.cshtml
@@ -1,5 +1,5 @@
 @page
-@model WeatherApp.Pages.Weather.IndexModel
+@model WeatherApp.Pages.WeatherView.IndexModel
 
 @{
     ViewData["Title"] = "Index";
