+++ WeatherApp/Pages/WeatherView/Index.cshtml.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 using WeatherApp.Models;
 
-namespace WeatherApp.Pages.Weather
+namespace WeatherApp.Pages.WeatherView
 {
     public class IndexModel : PageModel
     {
